requirements.txt
BeautifulSoup==3.2.1
Django==1.5
Fabric==1.6.0
Paste==1.7.5.1
Pillow==2.1.0
PyYAML==3.10
SQLAlchemy==0.8.1
South==0.7.6
WebOb==1.2.3
amqp==1.0.11
anyjson==0.3.3
argparse==1.2.1
astroid==1.0.0
billiard==2.7.3.28
celery==3.0.19
distribute==0.6.49
django-addendum==0.1.0
django-announce==0.1.5
django-appconf==0.6
django-avatar==1.0.5
django-blog-zinnia==0.12.3
django-celery==3.0.17
django-celery-email==1.0.3
django-classy-tags==0.4
django-compressor==1.3
django-debug-toolbar==0.9.4
django-debug-toolbar-template-timings==0.5.1
django-grappelli==2.4.6
django-jsonfield==0.9.3
django-mobile==0.3.0
django-mptt==0.5.5
django-postgrespool==0.2.4
django-pure-pagination==0.2.1
django-redis-sessions==0.4.0
django-sendgrid==1.0.1
django-tagging==0.3.1
django-templated-email==0.4.9
django-timezone-field==0.4
django-tinymce==1.5.1
django-tracking==0.4.1
django-watson==1.1.1
django-xmlrpc==0.1.5
djangorestframework==2.3.6
eventlet==0.14.0
gdata==2.0.17
google-api-python-client==1.1
greenlet==0.4.1
gunicorn==0.17.2
httplib2==0.8
inspector-panel==0.0.1
ipython==1.1.0
iso8601==0.1.4
kombu==2.5.10
localtunnel==0.6.1
logilab-common==0.60.0
lxml==3.2.3
newrelic==1.11.0.55
numpy==1.7.1
oauth2==1.5.211
oauthlib==0.4.0
paramiko==1.10.0
psycopg2==2.4.6
pycrypto==2.6
pylint==1.0.0
pyparsing==1.5.7
python-dateutil==2.1
python-gflags==2.0
python-memcached==1.48
python-openid==2.2.5
python-simple-hipchat==0.1
python-varnish==0.2.1
pytz==2013b
raven==3.3.6
redis==2.7.2
requests==1.2.0
requests-oauthlib==0.3.0
six==1.3.0
sockjs-tornado==1.0.0
tornado==3.1
tornado-redis==2.4.1
webapp2==2.5.2
wsgiref==0.1.2
zinnia-theme-bootstrap==0.1
